pyproject.toml
[project]
name = "portrait-analyser"
version = "0.3.0"
description = "Python module for portrait picture analysis"
authors = [
    {name = "Michał Pasternak", email = "michal.dtz@gmail.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "piexif>=1.1.3",
    "pillow>=10.2.0",
    "numpy>=1.26.4",
    "mediapipe>=0.10.0",
    "pyheif-iplweb>=0.7.1.dev1176; sys_platform == 'darwin'",
    "pyheif>=0.7.1; sys_platform == 'linux'",
]

[project.optional-dependencies]
pose = ["mediapipe>=0.10.30"]

[dependency-groups]
dev = [
    "pytest>=8.1.1",
]

[project.scripts]
analyse-portrait = "portrait_analyser.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "src"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
